Addded support for a LoginException in TorrentSearchProvider. * Updated various private tracker adapters to toss a LoginException when invalid login credentials are provided.
